Remove pagination, remove empty resources, update to questions
