+++ back/admin/resources/models.py
@@ -121,10 +121,7 @@ class Resource(BaseItem):
 
     @property
     def first_chapter_id(self):
-        try:
-            return self.chapters.all()[0].id
-        except:  # noqa E722
-            return 0
+        return self.chapters.all()[0].id
 
     def next_chapter(self, current_id, course):
         # We can't fetch course from the object, as the user might have already
